Finally changed the password logic.
1) Now the program starts up in unauthenticated mode (even if theres a
previously encrypted secret in the ini file - it cannot be decrypted if
it does not know the password). The default "fffuuuuuu" password was
removed, and now the password specified in the "Password" TAB must be
specified and the "DeCrypt" button clicked in order to use authenticated
mode. (No more need for an app restart, wrote code to do a simple
socket/client reconnect  with the newly decrypted secret).  Future
changes may simplify this further, or approach a compromise between a
hard-coded password (before), and a non-saved password required to be
entered every launch(now)
2) Fixed the code that inserts order-Id's into the EditBox when it
receives data over the socket. Previously the behavior was to insert the
ID of every OID transmitted over the socket, and it was grabbing
previous orders that have become fully active or portions active, since
a new order was added that changed the funded/unfunded status of
previous orders. Now the edit-box is only populated if the order-status
is "Post-Pending" meaning (Just-added), so it only receives the most
recent order.
3) The user orders box is now sortable. Defaults to sorting by price.
Also, I had problems in setting the columns to the right size, so I was
forced to use dummy data so resizeColumnsToContents() actually has some
data to work-off-of .
4) Some changes to goxapi, i have yet to remove the direct link of the
link to prof7bits commit and replace it with my own file, I will do tihs
later.