Reverse behavior. Ask = Buy, Bid = Sell.
Works out better for people who fulfill other orders (and meshes with
the behavior of the price button better, as it already does this).  For
people who are "copycats" and create orders based on similar orders on
the same side rather than fulfill orders, its going to take some getting
used to, as this is opposite of the original goxgui.