Add group_orders to preferences window.
Make the wallet use whatever multi-currency was specified from the
config file. Everything is working except JPY and SEK since they are the
only ones that use 3 decimals.  The money.py and currency.py is quite a
"large" and superfluous way to handle this. I should be able to do what
those lines of code do in much fewer.