Add TRAILING stop loss + Fix size&total 999999 issue
Trailing stop loss market sells can now be used, by setting price to a
negative number meaning the distance away from the current price to set
the stop loss to. (ie: -5, if current price is $115, stop loss price
will be $110. If the price goes up to $125, the stop loss trails behind
it as $120, and if the price now drops to $120 or below, we will take
profits and sell off at $120